pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Travel planning agent powered by LangGraph and DeepSeek"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-deepseek>=0.1.0",
    "langgraph>=0.2.0",
    "langchain-mcp-adapters>=0.2.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
backend = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
